requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
